Manual

AutoSymPTR.autosymptr — Method

autosymptr(f, B::AbstractMatrix, syms=nothing;
atol=0, rtol=sqrt(eps()), maxevals=typemax(Int64),
npt1=npt_update(f,0), npt2=npt_update(f,npt1),
rule1=ptr(npt1, f, B, syms), rule2=ptr(npt2, f, B, syms))

Computes the integral of f to within the specified tolerances and returns a tuple (I, E, numevals, rules) containing the estimated integral, the estimated error, the number of integrand evaluations, and a NamedTuple of rules used to compute the PTR on the most refined grid.

Convergence depends on periodicity

If the routine takes a long time to return, double check at the period of the function f along the basis vectors in the columns of B is consistent.

AutoSymPTR.symptr — Method

symptr(f, B::AbstractMatrix, syms=nothing;
npt=npt_update(f,0), rule=ptr(npt,f,B,syms))

Computes an npt symmetrized PTR rule of f on a unit hypercube in basis B and returns a tuple (I, rules) the integral, I, and the PTR rules used. Optionally uses a precomputed keyword rule.