Manual

AutoSymPTR.autoptr — Method

autoptr(f, B; kwargs...)

Same as autosymptr with trivial symmetries.

AutoSymPTR.autosymptr — Function

autosymptr(f, B::AbstractMatrix, syms; atol=0, rtol=sqrt(eps()), maxevals=typemax(Int64), buffer=nothing)

Computes the integral of f to within the specified tolerances and returns a tuple (I, E, numevals, rules) containing the estimated integral, the estimated error, the number of integrand evaluations. The integral on the symmetrized domain needs to be mapped back to the full domain by the caller, since for array-valued integrands this depends on the representation of the integrand under the action of the symmetries.

Convergence depends on periodicity

If the routine takes a long time to return, double check at the period of the function f along the basis vectors in the columns of B is consistent.

AutoSymPTR.ptr — Method

ptr(f, B::AbstractMatrix, syms; npt=npt_update(f,0), rule=PTRRule(float(eltype(B)),npt,Val(checksquare(B))))

Evaluates the npt^d point PTR rule on the integrand f. The coordinates are mapped into the basis B, whose basis vectors are stored as columns of the matrix. The integral is returned.

AutoSymPTR.symptr — Method

symptr(f, B::AbstractMatrix, syms; npt=npt_update(f,0), rule=symptr_rule(float(eltype(B)),npt,checksquare(B),syms))

Evaluates the npt^d point PTR rule on the integrand f. The coordinates are mapped into the basis B, whose basis vectors are stored as columns of the matrix. The integral is returned, and if the symmetries are nontrivial, the integral on the symmetry-reduced domain is returned, which the caller needs to remap to the full domain